Working ad filter that the user can enable and disable from UI. The Meddle server needs to know the address of the dns server that contains the list of domains to filter. In the forward path, if the user has enabled ad filtering the dns requests are redirected to our dns server that filters ads. In the reverse path, the source IP address of the DNS response is set to that of the DNS server
Former-commit-id: 00cd13e8d9e95857af9427dbbbfe5afbabc4d839
Former-commit-id: 308e85c213bc20ca78992dafa427e79fa5ee044a
Former-commit-id: 65cb21fc4e75b9cfc7d24553482c1c2553b5d425

+++ code/SimplePacketFilter/PktFilterModule/UserConfigs.cc
@@ -119,7 +119,7 @@ std::ostream& operator<<(std::ostream& os, const UserConfigs& userMapTable)
 	for(iter = userMapTable.idUserConfigMap.begin(); iter != userMapTable.idUserConfigMap.end(); iter++)
 	{
 		user_config_entry_t entry = iter->second;
-		os << "| " << entry.userID << " | " << entry.userName << " | " << entry.filterAdsAnalytics << std::endl;
+		os << "| " << entry.userID << " | " << entry.userName << " | " << (uint32_t)(entry.filterAdsAnalytics) << std::endl;
 	}
 	os << "-----------" << std::endl;
 	return os;
